requirements.txt
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.4.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
langdetect==1.0.9
numpy==2.2.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
urllib3==2.3.0
